Adicionando validação se categoria informada ja foi vinculada no momento da exclusão.

+++ Modulos/GerenciamentoMensal/Application/Categoria/Service/CategoriaService.cs
@@ -51,6 +51,9 @@ namespace Application.Implementacoes
             if (categoria is null)
                 return Result.Failure(Error.NotFound("Categoria informada não existe!"));
 
+            if (await _categoriaRepository.CategoriaJaFoiVinculado(categoria))
+                return Result.Failure(Error.Validation("Categoria informada possui vinculos realizados, por favor removas e após realize a exclusão!"));
+
             await _categoriaRepository.Delete(categoria);
 
             return Result.Success();
